T-273/T-269/T-274/T-275: DEBUG sets the day; the day's banner, note and trim
T-273. The clock's weekday is read with the hour (the S-3511's weekday
register; without a clock, one day per hour of play). DEBUG's TIME row: A
steps the watch, LEFT/RIGHT the weekday (AUTO, SUN..SAT), and the map is
relit at once. Played: "TIME NIGHT WED".

T-269. Out of a door, the popup that names the town slides in again with the
day: "F · WEDNESDAY · NIGHT" -- the day's note, the weekday, the watch -- in
the popup's own ink and paper, widened as a floor number widens it. DRAFT.
Played: out of the lab, "BLANCHE TOWN", then the day.

T-274. TUNE at the bedroom terminal sounds the day's note (SE_NOTE_C..B,
Sunday first) before its song.

T-275. The CHECKPOINT rotunda's trim (General row 2, index 8, which only it
draws) takes the day's colour as the palette loads, before the watch's tint;
seven colours from the user's rainbow, each 22+ CIE76 from every type hue
(DAEMONS tools/gbadaytrim.py -> src/data/day_trims.h). The primary rows'
signature now carries the weekday. Played: CALLOW's rotunda, SUN to SAT.

+++ include/fieldmap.h
@@ -63,7 +63,7 @@ void DaemonsReloadPrimaryTilesetPalette(const struct MapLayout *mapLayout);
 void DaemonsSetHalftoneColour(bool8 inColour);
 void DaemonsGreyHalftoneFrame(void);
 u8 DaemonsFieldTint(void);
-u8 DaemonsPaletteSignature(void);
-u8 DaemonsPrimarySignature(void);
+u16 DaemonsPaletteSignature(void);
+u16 DaemonsPrimarySignature(void);
 
 #endif //GUARD_FIELDMAP_H
